requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
SQLAlchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10
pydantic-settings==2.11.0
python-multipart==0.0.20
requests==2.32.5
pytz==2025.2
APScheduler==3.11.0

pydantic-settings==2.11.0
python-multipart==0.0.20
requests==2.32.5
pytz==2025.2  # For timezone handling
APScheduler==3.11.0
pytest>=8.3.3
pytest-asyncio>=0.21.0

# === HTTP Client for API Tests ===
httpx>=0.27.2

# === Environment Configuration ===
python-dotenv==1.1.1


# Core Testing Dependencies
pytest>=8.3.3
pytest-asyncio>=0.21.0
httpx>=0.27.2

# Environment Configuration
python-dotenv==1.1.1

# Data Processing and XGBoost Dependencies
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
scipy==1.16.3
six==1.17.0
tzdata==2025.2
xgboost==3.1.1

